requirements.txt
google-generativeai
python-dotenv
requests
feedparser
groq
gspread
oauth2client
openai
google-genai
